requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pandas==2.0.3
numpy==1.26.4
openpyxl==3.1.2
python-multipart==0.0.6
sqlalchemy
psycopg2-binary
pydantic
python-dotenv